test: move form-urlencoded example coverage to core
